Counting facility
Allow sections to count the items they created, forwarded etc., and
provide a method to print a summary.
This requires the blueprints to create their count methods themselves,
but this is done with a convinient factory function, and the counting
itself is designed to be as performant as possible.  The final summary
allows e.g. to spot the sections were many items unexpectedly went lost.

Conflicts:
	src/collective/transmogrifier/transmogrifier.py