itemInfo facility
A little development/debugging helper for pipelines: the itemInfo
functions print a compact information about the given item. The values
of a selected subset of keys are printed (_type and _path by default),
and the remaining keys are listed.

By default, only the first item in a loop is printed. It is possible to
explicitly show the first item with a certain quality (specify
showone='<key>' in the calling code).  The itemInfo function returns a
boolean value which tells if it decided to print.