Optional verbosity for ConfigurationRegistry.clear
When ConfigurationRegistry.clear is fired unexpectedly, this can go
unnoticed and cause transmogrifications to fail.  Thus, this can be
customized by setting the CLEANUP variable.

To suppress the call by the zope.testing.cleanup facility entirely, this
variable must must be changed to OFF in the module code, because this is
done on import time.  For other uses, the variable can be set after
importing.

The present SILENT value reflects the so-far behaviour.

Conflicts:
	src/collective/transmogrifier/transmogrifier.py